Refactor qb_taxijob: Modularize client scripts and add backup for legacy code
- Split main client script into multiple modular files for better organization and maintainability.
- Added a backup of the original monolithic client script as `backup_main_legacy.lua`.
- Introduced new files: `bookings.lua`, `garage.lua`, `init.lua`, `meter.lua`, `npc.lua`, and `state.lua` to handle specific functionalities.
- Updated `fxmanifest.lua` to include new client scripts.
- Improved code structure and readability across the new files.
